code was refactored & block creating added

+++ src/main/java/ru/mycubecraft/Settings.java
@@ -19,8 +19,8 @@ public final class Settings {
 
     public static final float MOUSE_SENSITIVITY = 0.095f;
 
-    public static final float MAX_LOOK = 120.0f;
-    public static final float MIN_LOOK = -120.0f;
+    public static final double MAX_LOOK = 120.00;
+    public static final double MIN_LOOK = -120.00;
 
     public static final float SKY_BOX_SCALE = 100;
